requirements.txt
PyYAML==3.11
pystache==0.5.4